Add 'Random' node types
