+++ addons/geoscript/types/geometry.py
@@ -10,7 +10,11 @@ from .boolean import Boolean
 class Geometry(AbstractSocket):
     """Corresponds to a Geometry socket type in Blender's Geometry Nodes"""
     
-    def __init__(self, node_tree: bpy.types.NodeTree = None, socket_reference: bpy.types.NodeSocket = None, layer: int = 0):
+    def __init__(
+            self,
+            node_tree: bpy.types.NodeTree = None,
+            socket_reference: bpy.types.NodeSocket = None,
+            layer: int = 0):
         super().__init__(node_tree, socket_reference, layer)
     
     @staticmethod
